invoke tcl stuff from python without metaprogramming
hat tip to andrew brooks for pointing out that the TclProc class can
define `__call__` and thus be callable as is.  By maintaining state
of whether the proc/command correspoing to the TclProc object is a
proc or a C command, the call function call invoke the right trampoline,
either the big one or the passthrough one, depending on which type of
function it is shadowing.