new ShadowDict "get" method
new "get" method for ShadowDict, takes optional named parameters "to"
and "default".

If the requested key is not found and a default is specified to "get", return the
default.  If the key isn't found and default isn't specified, return the
ShadowDict default defined at creation time, if one was defined, else
raise an error.

if "to" is specified to "get", a conversion to that datatype is attempted,
and an error is raised if the conversion fails.  If "to" was specified
at ShadowDict creation time, a conversion to that type is attempted,
else a conversion to str is attempted, which probably won't fail lol.

Even if default values are used, the to conversion is performed.