Rip out passing module pointer around
The tcl return functions had been changed to get the python tohil module
pointer, quite a few calls then needed to have that available.

It's so easy to get using PyImport_AddModule, and it doesn't happen
really except for handling errors, that we just do that when we need it
and otherwise don't pass that extra argument all over the place.