allow a callable function to be specified in a to= arg
If the to= argument is not a support python type but is callbale, it
will be called with one argument, a tclobj containing the object to be
returned.  It is up to this function to manipulate the object in some
way, and whatever the function returns is what the relevant tohil function will
return.