fix overwrites for tclobj methods when bound to vars
There were errors in how tclvars (tclobjs bound to vars) handled some methods such as insert, pop, append, and extend.

We introduce a new routine, TohilTclObj_objptr_for_write, that returns a
TclObj * that we can mutate.  It is either from the tclobj object
directly or from the tcl variable defined in the tclobj (tclvar), and
it's been pre-checked to see if it's shared and if it is, it's been
duplicated properly such that we can mutate it (it is no longer shared.)

We also subsume a couple places where tohil C code "got it right" but in
an ad hoc manner with calls to the new function.

fixes issue #38