load tcl via package require when python is the initiator
Previously we called Tohil_Init when initializing Tcl from the Python
side (in tohil_mod_exec), but while this loaded the tohil C extension
just fine, it did not load the Tcl source code that is also a part of
the package.

With this change, if Python is initializing Tcl, it now does so with a
package require rather than a Tohil_Init, solving that problem.

Now, any Python code that imports tohil, including subinterpreters, will
get all of tohil rather than just the C extension part.