make tohil::call support builtins
The Tcl-side tohil::call function provides a way to call a Python
function while making sure that the arguments passed to the function
are not interpreted by Python.

Because it didn't look in builtins if it couldn't find
the function in dunder main, stuff like "tohil::call print foo"
didn't work.

We now look in builtins after exhausting our other options.

Thanks to Oliver Jowett (obj) for a consult.

Fixes issue #39.