Don't error on tohil namespace creation failure
There's a report from FlightAware hitting this error, but it's elusive,
happens only in the rivet apache module and not very often.

But the error can be caused simply by doing a

```tcl
    namespace eval ::tohil {}
    package require tohil
```

That ought to not be an error.  In fact, from high-level Tcl itself there's no
explicit way to cause a namespace-already-exists error because `namespace eval` creates the namespace
if it doesn't exist and is ok with it if it does.

The concern here is that there is some underlying problem with the
multiple tcl interpreter / python subinterpreter stuff.  But in any case
the above shouldn't be a traceback.