import modules tohil needs as dunder names
The tohil module imports or imports from a number of modules, that then
show up in the output of dir(tohil).

This commit imports the modules with an "as" argument to import them
with a leading underscore, for example "import traceback" becomes
"import traceback as _traceback" and
"from collections.abc import MutableMapping" becomes
"from collections.abc import MutableMapping as _MutableMapping" and
changes all the references accordingly, so instead of referencing
"sys.stdout" we reference "_sys.stdout", which makes dir happy.

Fixes Issue #30