fix crash when deleting tcl interpreters
when a tcl child interpreter has been created with "interp create" and
then been told to do a "package require tohil" and then deleted, and
python was the parent, tohil would segfault.

these changes fix that.  we use a struct to maintain pointers for each
tcl interpreter that includes tohil, TohilPyterps, that has a pointer to
the parent python interpreter and the child python interpreter.  for the
parent interpreter, be it tcl or python, the parent is the same as the
child.

for all other tcl interpreters (ones created via interp create), the
parent is the parent tcl interpreter's python interpreter state, and the
child is a new python interpreter state that we created for it.

now when a child tcl interpreter is deleted, the callback function that
tohil registered to be called when that occurs can switch context to the
child python interpreter, delete it, then switch context to the parent
python interpreter.

this fixes the crash of:

```
import tohil

i = tohil.eval("interp create")

tohil.call(i, "eval", 'package require tohil')
tohil.call("interp", "delete", i)
```