stop stashing tcl interp ptr in the tclobj and tcldict types
Thanks to @nasageek for identifying this problem, which is setting a
static value in two data types that are used by multiple python
subinterpreters.  We simplify by getting rid of that and finding the tcl
interpreter the same way we do elsehwere when we need it and have been
given no direct handle to it.